cleaned up code (include physics)
